pyproject.toml
[project]
name = "wikpage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
    "tinyfish>=0.2.4",
]
